requirement.txt
openvas_lib==1.1.4
python-gvm==24.1.0
gvm-tools==24.1.0
customtkinter==5.2.2
pillow==10.2.0
reportlab==4.1.0
scapy==2.5.0
